OA-3 Tipos de dados gerados nas respostas de inteiro alterados. Long -> Integer (bug)

+++ otus-activity/src/main/java/org/ccem/otus/model/survey/activity/filling/answer/GridIntegerAnswer.java
@@ -3,13 +3,13 @@ package org.ccem.otus.model.survey.activity.filling.answer;
 public class GridIntegerAnswer {
 	private String objectType;
 	private String customID;
-	private Long value;
+	private Integer value;
 
 	public String getCustomID() {
 		return customID;
 	}
 
-	public Long getValue() {
+	public Integer getValue() {
 		return value;
 	}
